+++ Curso-Redux-por-Bedu/blog/src/actions/publicacionesActions.js
@@ -1,23 +1,23 @@
 import axios from 'axios';
-import { TRAER_TODOS, CARGANDO, ERROR } from '../types/publicacionesTypes';
+import {
+  CARGANDO,
+  ERROR,
+  TRAER_POR_USUARIO
+} from '../types/publicacionesTypes';
 
-export const traerTodos = () => async (dispatch) => {
-	dispatch({
-		type: CARGANDO
-	});
+export const traerPorUsuario = (key) => async (dispatch, getState) => {
+  const { usuarios } = getState().usuariosReducer;
+  const { publicaciones } = getState().publicacionesReducer;
+  const usuario_id = usuarios[key].id;
 
-	try {
-		const respuesta = await axios.get('https://jsonplaceholder.typicode.com/posts');
-		dispatch({
-			type: TRAER_TODOS,
-			payload: respuesta.data
-		})
-	}
-	catch (error) {
-		console.log(error.message);
-		dispatch({
-			type: ERROR,
-			payload: 'Algo salió mal, intente más tarde.'
-		})
-	}
+  const respuesta = await axios.get(`https://jsonplaceholder.typicode.com/posts?userId=${usuario_id}`);
+  const publicaciones_actualizadas = [
+    ...publicaciones,
+    respuesta.data
+  ];
+
+  dispatch({
+    type: TRAER_POR_USUARIO,
+    payload: publicaciones_actualizadas
+  });
 };
