+++ Curso-Redux-por-Bedu/blog/src/actions/usuariosActions.js
@@ -2,18 +2,22 @@ import axios from 'axios';
 import { TRAER_TODOS, CARGANDO, ERROR } from '../types/usuariosTypes';
 
 export const traerTodos = () => async (dispatch) => {
-  dispatch({ type: CARGANDO })
+  dispatch({
+    type: CARGANDO
+  });
+
   try {
     const respuesta = await axios.get('https://jsonplaceholder.typicode.com/users');
     dispatch({
       type: TRAER_TODOS,
       payload: respuesta.data
     })
-  } catch (err) {
-    console.log('Error:', err.message);
+  }
+  catch (error) {
+    console.log(error.message);
     dispatch({
-      type:ERROR,
-      payload: err.message
+      type: ERROR,
+      payload: 'Algo salió mal, intente más tarde.'
     })
   }
 };
