+++ Curso-Redux-por-Bedu/blog/src/components/Publicaciones/index.js
@@ -5,14 +5,15 @@ import * as usuariosActions from '../../actions/usuariosActions';
 import * as publicacionesActions from '../../actions/publicacionesActions';
 
 const { traerTodos: usuariosTraerTodos } = usuariosActions;
-const { traerTodos: publicacionesTraerTodos } = publicacionesActions;
+const { traerPorUsuario: publicacionesTraerPorUsuario } = publicacionesActions;
 
 class Publicaciones extends Component {
 
-  componentDidMount() {
+  async componentDidMount() {
     if (!this.props.usuariosReducer.usuarios.length) {
-      this.props.usuariosTraerTodos();
+      await this.props.usuariosTraerTodos();
     }
+    this.props.publicacionesTraerPorUsuario(this.props.match.params.key);
   }
 
   render() {
@@ -34,7 +35,7 @@ const mapStateToProps = ({ usuariosReducer, publicacionesReducer }) => {
 
 const mapDispatchToProps = {
   usuariosTraerTodos,
-  publicacionesTraerTodos
+  publicacionesTraerPorUsuario
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Publicaciones);
