+++ Curso-Redux-por-Bedu/blog/src/components/Usuarios/index.js
@@ -8,8 +8,10 @@ import * as usuariosActions from '../../actions/usuariosActions';
 
 class Usuarios extends Component {
 
-  async componentDidMount() {
-    this.props.traerTodos();
+  componentDidMount() {
+    if (!this.props.usuarios.length) {
+      this.props.traerTodos();
+    }
   }
 
   ponerContenido = () => {
@@ -25,6 +27,7 @@ class Usuarios extends Component {
   };
 
   render() {
+    console.log(this.props)
     return (
       <div>
         <h1>Usuarios</h1>
