+++ Curso-Redux-por-Bedu/blog/src/reducers/publicacionesReducer.js
@@ -1,4 +1,8 @@
-import { TRAER_TODOS, CARGANDO, ERROR } from '../types/publicacionesTypes';
+import {
+	CARGANDO,
+	ERROR,
+	TRAER_POR_USUARIO
+} from '../types/publicacionesTypes';
 
 const INITIAL_STATE = {
 	publicaciones: [],
@@ -8,7 +12,7 @@ const INITIAL_STATE = {
 
 export default (state = INITIAL_STATE, action) => {
 	switch (action.type) {
-		case TRAER_TODOS:
+		case TRAER_POR_USUARIO:
 			return {
 				...state,
 				publicaciones: action.payload,
