+++ Curso-Redux-por-Bedu/blog/src/types/publicacionesTypes.js
@@ -1,3 +1,3 @@
-export const TRAER_TODOS = 'publicaciones_traer_todos';
 export const CARGANDO = 'publicaciones_cargando';
-export const ERROR = 'publicaciones_ERROR';
+export const ERROR = 'publicaciones_error';
+export const TRAER_POR_USUARIO = 'publicaciones_traer_por_usuario';
